+++ src/lib.rs
@@ -201,9 +201,15 @@ pub async fn development_transport(
     keypair: identity::Keypair,
 ) -> std::io::Result<core::transport::Boxed<(PeerId, core::muxing::StreamMuxerBox)>> {
     let transport = {
-        let dns_tcp = dns::DnsConfig::system(tcp::TcpConfig::new().nodelay(true)).await?;
+        let dns_tcp = dns::DnsConfig::system(tcp::TcpTransport::new(
+            tcp::GenTcpConfig::new().nodelay(true),
+        ))
+        .await?;
         let ws_dns_tcp = websocket::WsConfig::new(
-            dns::DnsConfig::system(tcp::TcpConfig::new().nodelay(true)).await?,
+            dns::DnsConfig::system(tcp::TcpTransport::new(
+                tcp::GenTcpConfig::new().nodelay(true),
+            ))
+            .await?,
         );
         dns_tcp.or_transport(ws_dns_tcp)
     };
@@ -259,9 +265,11 @@ pub fn tokio_development_transport(
     keypair: identity::Keypair,
 ) -> std::io::Result<core::transport::Boxed<(PeerId, core::muxing::StreamMuxerBox)>> {
     let transport = {
-        let dns_tcp = dns::TokioDnsConfig::system(tcp::TokioTcpConfig::new().nodelay(true))?;
+        let dns_tcp = dns::TokioDnsConfig::system(tcp::TokioTcpTransport::new(
+            tcp::GenTcpConfig::new().nodelay(true),
+        ))?;
         let ws_dns_tcp = websocket::WsConfig::new(dns::TokioDnsConfig::system(
-            tcp::TokioTcpConfig::new().nodelay(true),
+            tcp::TokioTcpTransport::new(tcp::GenTcpConfig::new().nodelay(true)),
         )?);
         dns_tcp.or_transport(ws_dns_tcp)
     };