+++ protocols/relay/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - Do not duplicate the p2p/xxx component with the relay PeerId when a client requests a reservation. See [PR 2701].
 
+- Expose `HOP_PROTOCOL_NAME` and `STOP_PROTOCOL_NAME`.
+
 [PR 2701]: https://github.com/libp2p/rust-libp2p/pull/2701/
 
 # 0.9.1